Move checkpoint schedule code into tlm_adjoint/checkpoint_schedules
