+++ ACKNOWLEDGEMENTS
@@ -110,9 +110,10 @@ and adjoint information using the approach described in [1, 2].
 
 ===============================================================================
 
-tlm_adjoint implements the binomial checkpointing strategy described in [3].
-tlm_adjoint further implements multi-stage offline checkpointing, using a brute
-force approach to yield behaviour described in [4].
+The MultistageCheckpointSchedule class in
+tlm_adjoint/checkpoint_schedules/binomial.py implements the binomial
+checkpointing strategy described in [3], with multi-stage offline checkpointing
+schedules equivalent to those described in [4].
 
 [3]  A. Griewank and A. Walther, "Algorithm 799: Revolve: An implementation of
      checkpointing for the reverse or adjoint mode of computational
@@ -124,9 +125,9 @@ force approach to yield behaviour described in [4].
 
 ===============================================================================
 
-The TwoLevelCheckpointSchedule class in tlm_adjoint/binomial_checkpointing.py
-implements the two-level mixed periodic/binomial checkpointing approach
-described in [5, 6].
+The TwoLevelCheckpointSchedule class in
+tlm_adjoint/checkpoint_schedules/binomial.py implements the two-level mixed
+periodic/binomial checkpointing approach described in [5, 6].
 
 [5]  Gavin J. Pringle, Daniel C. Jones, Sudipta Goswami, Sri Hari Krishna
      Narayanan, and Daniel Goldberg, "Providing the ARCHER community with
