+++ tlm_adjoint/checkpoint_schedules/memory.py
@@ -1,0 +1,67 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+# For tlm_adjoint copyright information see ACKNOWLEDGEMENTS in the tlm_adjoint
+# root directory
+
+# This file is part of tlm_adjoint.
+#
+# tlm_adjoint is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, version 3 of the License.
+#
+# tlm_adjoint is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
+
+from .schedule import CheckpointSchedule, Configure, Forward, Reverse, \
+    EndForward, EndReverse
+
+import sys
+
+__all__ = \
+    [
+        "MemoryCheckpointSchedule"
+    ]
+
+
+class MemoryCheckpointSchedule(CheckpointSchedule):
+    def iter(self):
+        # Forward
+
+        if self._max_n is not None:
+            # Unexpected finalize
+            raise RuntimeError("Invalid checkpointing state")
+        yield Configure(True, True)
+
+        while self._max_n is None:
+            n0 = self._n
+            n1 = n0 + sys.maxsize
+            self._n = n1
+            yield Forward(n0, n1)
+
+        yield EndForward()
+
+        while True:
+            if self._r == 0:
+                # Reverse
+
+                self._r = self._max_n
+                yield Reverse(self._max_n, 0)
+            elif self._r == self._max_n:
+                # Reset for new reverse
+
+                self._r = 0
+                yield EndReverse(False,)
+            else:
+                raise RuntimeError("Invalid checkpointing state")
+
+    def is_exhausted(self):
+        return False
+
+    def uses_disk_storage(self):
+        return False
