+++ tlm_adjoint/checkpoint_schedules/periodic.py
@@ -1,0 +1,103 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+# For tlm_adjoint copyright information see ACKNOWLEDGEMENTS in the tlm_adjoint
+# root directory
+
+# This file is part of tlm_adjoint.
+#
+# tlm_adjoint is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, version 3 of the License.
+#
+# tlm_adjoint is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
+
+from .schedule import CheckpointSchedule, Clear, Configure, Forward, Reverse, \
+    Read, Write, EndForward, EndReverse
+
+__all__ = \
+    [
+        "PeriodicDiskCheckpointSchedule"
+    ]
+
+
+class PeriodicDiskCheckpointSchedule(CheckpointSchedule):
+    def __init__(self, period, *, keep_block_0_ics=False):
+        if period < 1:
+            raise ValueError("period must be positive")
+
+        super().__init__()
+        self._period = period
+        self._keep_block_0_ics = keep_block_0_ics
+
+    def iter(self):
+        # Forward
+
+        while self._max_n is None:
+            yield Configure(True, False)
+            if self._max_n is not None:
+                # Unexpected finalize
+                raise RuntimeError("Invalid checkpointing state")
+            n0 = self._n
+            n1 = n0 + self._period
+            self._n = n1
+            yield Forward(n0, n1)
+
+            # Finalize permitted here
+
+            yield Write(n0, "disk")
+            yield Clear(True, True)
+
+        yield EndForward()
+
+        while True:
+            # Reverse
+
+            while self._r < self._max_n:
+                n = self._max_n - self._r - 1
+                n0 = (n // self._period) * self._period
+                del n
+                n1 = min(n0 + self._period, self._max_n)
+                if self._r != self._max_n - n1:
+                    raise RuntimeError("Invalid checkpointing state")
+
+                self._n = n0
+                yield Read(n0, "disk", False)
+                yield Clear(True, True)
+
+                if self._keep_block_0_ics and n0 == 0:
+                    yield Configure(True, True)
+                    self._n = n0 + 1
+                    yield Forward(n0, n0 + 1)
+
+                    if n1 > n0 + 1:
+                        yield Configure(False, True)
+                        self._n = n1
+                        yield Forward(n0 + 1, n1)
+                else:
+                    yield Configure(False, True)
+                    self._n = n1
+                    yield Forward(n0, n1)
+
+                self._r = self._max_n - n0
+                yield Reverse(n1, n0)
+                yield Clear(not self._keep_block_0_ics or n0 != 0, True)
+            if self._r != self._max_n:
+                raise RuntimeError("Invalid checkpointing state")
+
+            # Reset for new reverse
+
+            self._r = 0
+            yield EndReverse(False,)
+
+    def is_exhausted(self):
+        return False
+
+    def uses_disk_storage(self):
+        return True
