+++ tlm_adjoint/checkpointing.py
@@ -25,16 +25,13 @@ from .interface import DEFAULT_COMM, function_copy, function_get_values, \
 
 from abc import ABC, abstractmethod
 from collections import deque
-import functools
 try:
     import mpi4py.MPI as MPI
 except ImportError:
     MPI = None
-import logging
 import numpy as np
 import os
 import pickle
-import sys
 import weakref
 
 __all__ = \
@@ -44,23 +41,7 @@ __all__ = \
 
         "Checkpoints",
         "PickleCheckpoints",
-        "HDF5Checkpoints",
-
-        "CheckpointAction",
-        "Clear",
-        "Configure",
-        "Forward",
-        "Reverse",
-        "Read",
-        "Write",
-        "EndForward",
-        "EndReverse",
-
-        "CheckpointSchedule",
-        "HRevolveCheckpointSchedule",
-        "NoneCheckpointSchedule",
-        "MemoryCheckpointSchedule",
-        "PeriodicDiskCheckpointSchedule"
+        "HDF5Checkpoints"
     ]
 
 
@@ -763,510 +744,3 @@ class HDF5Checkpoints(Checkpoints):
         self._comm.barrier()
         del self._cp_filenames[n]
         del self._cp_spaces[n]
-
-
-class CheckpointAction:
-    def __init__(self, *args):
-        self.args = args
-
-    def __repr__(self):
-        return f"{type(self).__name__}{self.args!r}"
-
-    def __len__(self):
-        """The number of forward or reverse steps evaluated by this action."""
-        return 0
-
-    def __eq__(self, other):
-        return type(self) == type(other) and self.args == other.args
-
-
-class Clear(CheckpointAction):
-    def __init__(self, clear_ics, clear_data):
-        super().__init__(clear_ics, clear_data)
-
-    @property
-    def clear_ics(self):
-        return self.args[0]
-
-    @property
-    def clear_data(self):
-        return self.args[1]
-
-
-class Configure(CheckpointAction):
-    def __init__(self, store_ics, store_data):
-        super().__init__(store_ics, store_data)
-
-    @property
-    def store_ics(self):
-        return self.args[0]
-
-    @property
-    def store_data(self):
-        return self.args[1]
-
-
-class Forward(CheckpointAction):
-    def __init__(self, n0, n1):
-        super().__init__(n0, n1)
-
-    def __iter__(self):
-        yield from range(self.n0, self.n1)
-
-    def __len__(self):
-        return self.n1 - self.n0
-
-    def __contains__(self, step):
-        return self.n0 <= step < self.n1
-
-    @property
-    def n0(self):
-        return self.args[0]
-
-    @property
-    def n1(self):
-        return self.args[1]
-
-
-class Reverse(CheckpointAction):
-    def __init__(self, n1, n0):
-        super().__init__(n1, n0)
-
-    def __iter__(self):
-        yield from range(self.n1 - 1, self.n0 - 1, -1)
-
-    def __len__(self):
-        return self.n1 - self.n0
-
-    def __contains__(self, step):
-        return self.n0 <= step < self.n1
-
-    @property
-    def n0(self):
-        return self.args[1]
-
-    @property
-    def n1(self):
-        return self.args[0]
-
-
-class Read(CheckpointAction):
-    def __init__(self, n, storage, delete):
-        super().__init__(n, storage, delete)
-
-    @property
-    def n(self):
-        return self.args[0]
-
-    @property
-    def storage(self):
-        return self.args[1]
-
-    @property
-    def delete(self):
-        return self.args[2]
-
-
-class Write(CheckpointAction):
-    def __init__(self, n, storage):
-        super().__init__(n, storage)
-
-    @property
-    def n(self):
-        return self.args[0]
-
-    @property
-    def storage(self):
-        return self.args[1]
-
-
-class EndForward(CheckpointAction):
-    pass
-
-
-class EndReverse(CheckpointAction):
-    def __init__(self, exhausted):
-        super().__init__(exhausted)
-
-    @property
-    def exhausted(self):
-        return self.args[0]
-
-
-class CheckpointSchedule(ABC):
-    """
-    A checkpointing schedule.
-
-    The schedule is defined by iter, which yields actions in a similar manner
-    to the approach used in
-       A. Griewank and A. Walther, "Algorithm 799: Revolve: An implementation
-       of checkpointing for the reverse or adjoint mode of computational
-       differentiation", ACM Transactions on Mathematical Software, 26(1), pp.
-       19--45, 2000
-    e.g. 'forward', 'read', and 'write' correspond to ADVANCE, RESTORE, and
-    TAKESHOT respectively in Griewank and Walther 2000 (although here 'write'
-    actions occur *after* forward advancement from snapshots).
-
-    The iter method yields (action, data), with:
-
-    Clear(clear_ics, clear_data)
-    Clear checkpoint storage. clear_ics indicates whether stored initial
-    condition data should be cleared. clear_data indicates whether stored
-    non-linear dependency data should be cleared.
-
-    Configure(store_ics, store_data)
-    Configure checkpoint storage. store_ics indicates whether initial condition
-    data should be stored. store_data indicates whether non-linear dependency
-    data should be stored.
-
-    Forward(n0, n1)
-    Run the forward from the start of block n0 to the start of block n1.
-
-    Reverse(n1, n0)
-    Run the adjoint from the start of block n1 to the start of block n0.
-
-    Read(n, storage, delete)
-    Read checkpoint data associated with block n from the indicated storage.
-    delete indicates whether the checkpoint data should be deleted.
-
-    Write(n, storage)
-    Write checkpoint data associated with block n to the indicated storage.
-
-    EndForward()
-    End the forward calculation.
-
-    EndReverse(exhausted)
-    End a reverse calculation. If exhausted is False then a further reverse
-    calculation can be performed.
-    """
-
-    def __init__(self, max_n=None):
-        if max_n is not None and max_n < 1:
-            raise ValueError("max_n must be positive")
-
-        self._n = 0
-        self._r = 0
-        self._max_n = max_n
-
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-
-        cls_iter = cls.iter
-
-        @functools.wraps(cls_iter)
-        def iter(self):
-            if not hasattr(self, "_iter"):
-                self._iter = cls_iter(self)
-            return self._iter
-
-        cls.iter = iter
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        return next(self.iter())
-
-    @abstractmethod
-    def iter(self):
-        raise NotImplementedError
-
-    @abstractmethod
-    def is_exhausted(self):
-        raise NotImplementedError
-
-    @abstractmethod
-    def uses_disk_storage(self):
-        raise NotImplementedError
-
-    def n(self):
-        return self._n
-
-    def r(self):
-        return self._r
-
-    def max_n(self):
-        return self._max_n
-
-    def is_running(self):
-        return hasattr(self, "_iter")
-
-    def finalize(self, n):
-        if n < 1:
-            raise ValueError("n must be positive")
-        if self._max_n is None:
-            if self._n >= n:
-                self._n = n
-                self._max_n = n
-            else:
-                raise RuntimeError("Invalid checkpointing state")
-        elif self._n != n or self._max_n != n:
-            raise RuntimeError("Invalid checkpointing state")
-
-
-class NoneCheckpointSchedule(CheckpointSchedule):
-    def iter(self):
-        # Forward
-
-        while self._max_n is None:
-            n0 = self._n
-            n1 = n0 + sys.maxsize
-            self._n = n1
-            yield Forward(n0, n1)
-
-        yield EndForward()
-
-    def is_exhausted(self):
-        return self._max_n is not None
-
-    def uses_disk_storage(self):
-        return False
-
-
-class MemoryCheckpointSchedule(CheckpointSchedule):
-    def iter(self):
-        # Forward
-
-        if self._max_n is not None:
-            # Unexpected finalize
-            raise RuntimeError("Invalid checkpointing state")
-        yield Configure(True, True)
-
-        while self._max_n is None:
-            n0 = self._n
-            n1 = n0 + sys.maxsize
-            self._n = n1
-            yield Forward(n0, n1)
-
-        yield EndForward()
-
-        while True:
-            if self._r == 0:
-                # Reverse
-
-                self._r = self._max_n
-                yield Reverse(self._max_n, 0)
-            elif self._r == self._max_n:
-                # Reset for new reverse
-
-                self._r = 0
-                yield EndReverse(False,)
-            else:
-                raise RuntimeError("Invalid checkpointing state")
-
-    def is_exhausted(self):
-        return False
-
-    def uses_disk_storage(self):
-        return False
-
-
-class PeriodicDiskCheckpointSchedule(CheckpointSchedule):
-    def __init__(self, period, *, keep_block_0_ics=False):
-        if period < 1:
-            raise ValueError("period must be positive")
-
-        super().__init__()
-        self._period = period
-        self._keep_block_0_ics = keep_block_0_ics
-
-    def iter(self):
-        # Forward
-
-        while self._max_n is None:
-            yield Configure(True, False)
-            if self._max_n is not None:
-                # Unexpected finalize
-                raise RuntimeError("Invalid checkpointing state")
-            n0 = self._n
-            n1 = n0 + self._period
-            self._n = n1
-            yield Forward(n0, n1)
-
-            # Finalize permitted here
-
-            yield Write(n0, "disk")
-            yield Clear(True, True)
-
-        yield EndForward()
-
-        while True:
-            # Reverse
-
-            while self._r < self._max_n:
-                n = self._max_n - self._r - 1
-                n0 = (n // self._period) * self._period
-                del n
-                n1 = min(n0 + self._period, self._max_n)
-                if self._r != self._max_n - n1:
-                    raise RuntimeError("Invalid checkpointing state")
-
-                self._n = n0
-                yield Read(n0, "disk", False)
-                yield Clear(True, True)
-
-                if self._keep_block_0_ics and n0 == 0:
-                    yield Configure(True, True)
-                    self._n = n0 + 1
-                    yield Forward(n0, n0 + 1)
-
-                    if n1 > n0 + 1:
-                        yield Configure(False, True)
-                        self._n = n1
-                        yield Forward(n0 + 1, n1)
-                else:
-                    yield Configure(False, True)
-                    self._n = n1
-                    yield Forward(n0, n1)
-
-                self._r = self._max_n - n0
-                yield Reverse(n1, n0)
-                yield Clear(not self._keep_block_0_ics or n0 != 0, True)
-            if self._r != self._max_n:
-                raise RuntimeError("Invalid checkpointing state")
-
-            # Reset for new reverse
-
-            self._r = 0
-            yield EndReverse(False,)
-
-    def is_exhausted(self):
-        return False
-
-    def uses_disk_storage(self):
-        return True
-
-
-class HRevolveCheckpointSchedule(CheckpointSchedule):
-    def __init__(self, max_n, snapshots_in_ram, snapshots_on_disk, *,
-                 keep_block_0_ics=False,
-                 wvect=(0.0, 0.1), rvect=(0.0, 0.1), uf=1.0, ub=2.0, **kwargs):
-        super().__init__(max_n)
-        self._snapshots_in_ram = snapshots_in_ram
-        self._snapshots_on_disk = snapshots_on_disk
-        self._exhausted = False
-        self._keep_block_0_ics = keep_block_0_ics
-
-        cvect = (snapshots_in_ram, snapshots_on_disk)
-        import hrevolve
-        schedule = hrevolve.hrevolve(max_n - 1, cvect, wvect, rvect,
-                                     uf=uf, ub=ub, **kwargs)
-        self._schedule = list(schedule)
-
-        logger = logging.getLogger("tlm_adjoint.checkpointing")
-        logger.debug(f"H-Revolve schedule: {str(self._schedule):s}")
-
-    def iter(self):
-        def action(i):
-            assert i >= 0 and i < len(self._schedule)
-            action = self._schedule[i]
-            cp_action = action.type
-            if cp_action == "Forward":
-                n_0 = action.index
-                n_1 = n_0 + 1
-                storage = None
-            elif cp_action == "Forwards":
-                cp_action = "Forward"
-                n_0, n_1 = action.index
-                n_1 += 1
-                storage = None
-            elif cp_action == "Backward":
-                n_0 = action.index
-                n_1 = None
-                storage = None
-            elif cp_action in ["Read", "Write", "Discard"]:
-                storage, n_0 = action.index
-                n_1 = None
-                storage = {0: "RAM", 1: "disk"}[storage]
-            else:
-                raise RuntimeError(f"Unexpected action: {cp_action:s}")
-            return cp_action, (n_0, n_1, storage)
-
-        snapshots = set()
-        deferred_cp = None
-        for i in range(len(self._schedule)):
-            cp_action, (n_0, n_1, storage) = action(i)
-
-            if cp_action == "Forward":
-                if n_0 != self._n:
-                    raise RuntimeError("Invalid checkpointing state")
-
-                yield Clear(True, True)
-                yield Configure(n_0 not in snapshots, False)
-                self._n = n_1
-                yield Forward(n_0, n_1)
-            elif cp_action == "Backward":
-                if n_0 != self._n:
-                    raise RuntimeError("Invalid checkpointing state")
-                if n_0 != self._max_n - self._r - 1:
-                    raise RuntimeError("Invalid checkpointing state")
-
-                if deferred_cp is not None:
-                    snapshots.add(deferred_cp[0])
-                    yield Write(*deferred_cp)
-                    deferred_cp = None
-
-                yield Clear(True, True)
-                yield Configure(self._keep_block_0_ics and n_0 == 0, True)
-                self._n = n_0 + 1
-                yield Forward(n_0, n_0 + 1)
-                if self._n == self._max_n:
-                    if self._r != 0:
-                        raise RuntimeError("Invalid checkpointing state")
-                    yield EndForward()
-                self._r += 1
-                yield Reverse(n_0 + 1, n_0)
-            elif cp_action == "Read":
-                if deferred_cp is not None:
-                    raise RuntimeError("Invalid checkpointing state")
-
-                if n_0 == self._max_n - self._r - 1:
-                    cp_delete = True
-                elif i < len(self._schedule) - 2:
-                    d_cp_action, (d_n_0, _, d_storage) = action(i + 2)
-                    if d_cp_action == "Discard":
-                        if d_n_0 != n_0 or d_storage != storage:
-                            raise RuntimeError("Invalid schedule")
-                        cp_delete = True
-                    else:
-                        cp_delete = False
-
-                yield Clear(True, True)
-                if cp_delete:
-                    snapshots.remove(n_0)
-                self._n = n_0
-                yield Read(n_0, storage, cp_delete)
-            elif cp_action == "Write":
-                if n_0 != self._n:
-                    raise RuntimeError("Invalid checkpointing state")
-
-                if deferred_cp is not None:
-                    snapshots.add(deferred_cp[0])
-                    yield Write(*deferred_cp)
-                    deferred_cp = None
-
-                deferred_cp = (n_0, storage)
-            elif cp_action == "Discard":
-                if i < 2:
-                    raise RuntimeError("Invalid schedule")
-                r_cp_action, (r_n_0, _, r_storage) = action(i - 2)
-                if r_cp_action != "Read" \
-                        or r_n_0 != n_0 \
-                        or r_storage != storage:
-                    raise RuntimeError("Invalid schedule")
-            else:
-                raise RuntimeError(f"Unexpected action: {cp_action:s}")
-
-        yield Clear(not self._keep_block_0_ics or self._n != 1, True)
-
-        if len(snapshots) != 0:
-            raise RuntimeError("Invalid checkpointing state")
-
-        self._exhausted = True
-        yield EndReverse(True,)
-
-    def is_exhausted(self):
-        return self._exhausted
-
-    def uses_disk_storage(self):
-        return self._snapshots_on_disk > 0
